V4.1 and V4.2
Created the “Create Questions Setup” page, implemented validation for the questions metadata input and created the user interface for the sandbox for creating questions.
Adjusted truth table generation script so we can take the truth table string for storing answers in the database. Created Questions table in database and added code to submit the questions created to this table.