V4.6
Added icons and configured the User Interface of the Sandbox. Modified quit button to a cross rather than a button which says “Quit”. Modified the Questions selection to show the questions you’ve already answered in green with the score at the bottom of the list. Also, allowed the user to press anywhere on the question cell to load it.